refactoring for summary component
