+++ src/components/ResultFiltersAndShare/ResultFiltersAndShare.tsx
@@ -1,8 +1,7 @@
 import React, { useContext } from 'react';
 import QuizContext from '../CioQuiz/context';
 import ShareButton from '../ShareButton/ShareButton';
-import { MATCHED_OPTIONS_PLACEHOLDER } from '../../constants';
-import { formatMatchedOptions } from '../../utils';
+import QuizResultsSummary from '../QuizResultsSummary/QuizResultsSummary';
 
 interface ResultFiltersAndShareProps {
   onShare: () => void;
@@ -16,31 +15,13 @@ function ResultFiltersAndShare({
   showShareButton,
 }: ResultFiltersAndShareProps) {
   const { state } = useContext(QuizContext);
-  const matchedOptions = state?.quiz?.matchedOptions || [];
-  const {
-    text = '',
-    is_active: isActive,
-    items_separator: itemsSeparator,
-    last_separator: lastSeparator,
-  } = state?.quiz?.resultsConfig?.desktop?.response_summary || {};
-  const isActiveSummary = !!isActive && !!matchedOptions.length && !!text?.length;
-  const [summaryFirstPart, summaryLastPart] = text?.split(MATCHED_OPTIONS_PLACEHOLDER) || [];
-  const matchedOptionsTemplate = formatMatchedOptions(
-    matchedOptions,
-    itemsSeparator,
-    lastSeparator
-  );
+  const matchedOptions = state?.quiz?.matchedOptions;
+  const summary = state?.quiz?.resultsConfig?.desktop?.response_summary ?? null;
 
   return (
     <div className='cio-results-filter-and-redo-container cio-results-button-group'>
       <div className='cio-results-filter-container'>
-        {isActiveSummary && (
-          <p className='cio-results-explanation'>
-            {!!summaryFirstPart?.length && summaryFirstPart}
-            <span>{matchedOptionsTemplate}</span>
-            {!!summaryLastPart?.length && summaryLastPart}
-          </p>
-        )}
+        <QuizResultsSummary summary={summary} matchedOptions={matchedOptions} />
       </div>
       <div className='cio-results-number-and-share-button-group'>
         {numberOfResults} {numberOfResults === 1 ? 'result' : 'results'}
